requirements.txt
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.1.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pefile==2023.2.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
